cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-api-python-client==2.*
google-auth==2.*
google-cloud-secret-manager==2.*
